requirements/dev.txt
# requirements/dev.txt
-r base.txt
django-debug-toolbar==4.2.0
black==23.12.1
flake8==6.1.0
isort==5.13.2